+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add logging of actions of Admin Client API ( restart clients, turn off clients, modify time, etc )
 
 ## [4.3.0] [r22.08]
 ### Added